[refactor] : 메서드의 역할에 따른 이동
